newcommit

+++ print_u_d_i.c
@@ -6,7 +6,7 @@
 /*   By: nbarreir <nbarreir@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/10 23:52:18 by csantos-          #+#    #+#             */
-/*   Updated: 2021/04/15 02:10:57 by nbarreir         ###   ########.fr       */
+/*   Updated: 2021/04/15 02:44:48 by nbarreir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ void			print_conferir_d_i(t_flags *flags, int size)
 ** Prints decimal and int
 */
 
-void			print_d_i(t_flags *flags, int num)
+void			print_d_i(t_flags *flags, long int num)
 {
 	int		size;
 	char	*number;
